Refactor differential to incremental
