+++ coral-incremental/src/test/resources/hive.xml
@@ -1,11 +1,18 @@
+<!--
+
+    Copyright 2023 LinkedIn Corporation. All rights reserved.
+    Licensed under the BSD-2 Clause license.
+    See LICENSE in the project root for license information.
+
+-->
 <configuration>
   <property>
     <name>hive.exec.scratchdir</name>
-    <value>${coral.differential.test.dir}/hive-scratch-dir</value>
+    <value>${coral.incremental.test.dir}/hive-scratch-dir</value>
   </property>
   <property>
     <name>hive.metastore.warehouse.dir</name>
-    <value>${coral.differential.test.dir}/warehouse</value>
+    <value>${coral.incremental.test.dir}/warehouse</value>
   </property>
   <property>
     <name>javax.jdo.option.ConnectionDriverName</name>
@@ -13,7 +20,7 @@
   </property>
   <property>
     <name>javax.jdo.option.ConnectionURL</name>
-    <value>jdbc:derby:;databaseName=${coral.differential.test.dir}/metastore.db;create=true</value>
+    <value>jdbc:derby:;databaseName=${coral.incremental.test.dir}/metastore.db;create=true</value>
   </property>
   <property>
     <name>hive.metastore.schema.verification</name>
