+++ coral-service/build.gradle
@@ -24,8 +24,8 @@ configurations.all {
 
 dependencies {
   implementation project(':coral-common')
-  implementation project(':coral-differential')
   implementation project(':coral-hive')
+  implementation project(':coral-incremental')
   implementation project(':coral-trino')
   implementation project(':coral-spark')
